New tests for compact picker, removed hex color refs
Major:
- Added parameterised tests of pure colour adjustments of compact view
- Added Activity tests for Compact View Activity
- Added Main Activity test
- Removed all references to hex in Views and used ints instead

Minor:
- Added buttons to show examples of square and round dialogues (not yet implemented)
- Made getIthColor() in AbstractColorPicker static so it can be used in other classes without an instance
- Increased precision in some places by using Doubles instead of Floats
- Added FrameLayout around menu in CompactColorPicker to listen to clicks
- Added tag to ImageView so that the color is saved to the View

+++ library/src/main/java/com/gerardbradshaw/library/ColorSlider.kt
@@ -20,7 +20,7 @@ class ColorSlider : FrameLayout {
 
   // ------------------------ PROPERTIES ------------------------
 
-  val max = context.resources.getInteger(R.integer.spectrum_color_count).toFloat()
+  val max = context.resources.getInteger(R.integer.spectrum_color_count).toDouble()
   private lateinit var seekBar: SeekBar
   private lateinit var gradientBar: FrameLayout
   private var listener: OnProgressChangedListener? = null
